:recycle: Restructuring applying IDD approach to the main src folder.
