+++ src/main/java/com/codurance/allaboard/App.java
@@ -5,10 +5,10 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.web.servlet.error.ErrorMvcAutoConfiguration;
 
 @SpringBootApplication(exclude = ErrorMvcAutoConfiguration.class)
-public class AllABoardApplication {
+public class App {
 
   public static void main(String[] args) {
-    SpringApplication.run(AllABoardApplication.class, args);
+    SpringApplication.run(App.class, args);
   }
 
 }
