+++ src/test/java/com/codurance/allaboard/acceptance/SaveSurveyFeature.java
@@ -3,9 +3,9 @@ package com.codurance.allaboard.acceptance;
 import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.verify;
 
-import com.codurance.allaboard.survey.infrastructure.repositories.Surveys;
-import com.codurance.allaboard.survey.model.Survey;
-import com.codurance.allaboard.survey.services.SaveSurvey;
+import com.codurance.allaboard.core.model.survey.Surveys;
+import com.codurance.allaboard.core.model.survey.Survey;
+import com.codurance.allaboard.core.actions.survey.SaveSurvey;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
