+++ src/test/java/com/codurance/allaboard/unit/GoogleTokenInterceptorShould.java
@@ -7,7 +7,7 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 
-import com.codurance.allaboard.config.security.interceptors.GoogleTokenInterceptor;
+import com.codurance.allaboard.web.infrastructure.interceptors.token.GoogleTokenInterceptor;
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdToken;
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdToken.Payload;
 import com.google.api.client.googleapis.auth.oauth2.GoogleIdTokenVerifier;
